// frontend/src/api/generated/models/automationRuleRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable issue tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */
import type { AutomationTrigger } from './automationTrigger';
import type { RuleActions } from './ruleActions';
import type { RuleConditions } from './ruleConditions';
import type { UserPublic } from './userPublic';

export interface AutomationRuleRead {
  id: number;
  team_id: number;
  name: string;
  trigger: AutomationTrigger;
  is_enabled: boolean;
  conditions: RuleConditions;
  actions: RuleActions;
  created_by: UserPublic;
  created_at: string;
  updated_at: string;
}